Testing models per files.
 GB and NN are using Jacks additional preprocessing with outlier removal. 
Able to improve on GB and NN.
Using stacked models did not have much of an improvement, but able to incorporate NN into it.

Additionally, evaluating the model takes time. Able to skip evaluation and use kaggle submissions to get a logrmse value with their true values.
Maybe we can create an api call to get instant results?