Finished Models with a Full Pipeline implementation
Created a Full Pipeline runnable with arguments. See the "readme" on what the pipeline allows. Added additional model tuning and code updates on all 3 files for model implementation to integrate with the full Pipeline. Note that the pipeline by default runs the GB and NN model with the evaluation function, with an average runtime of 10 minutes.